+++ examples/helloworld/src/helloworld.rs
@@ -180,13 +180,16 @@ pub struct ResponseResult {
 
 /// `Greeter` service definition
 pub struct Greeter;
+
 impl ::ntex_grpc::ServiceDef for Greeter {
     const NAME: &'static str = "helloworld.Greeter";
     type Methods = GreeterMethods;
 }
+
 pub enum GreeterMethods {
     SayHello(GreeterSayHelloMethod),
 }
+
 impl ::ntex_grpc::MethodsDef for GreeterMethods {
     #[inline]
     fn by_name(name: &str) -> Option<Self> {
@@ -197,49 +200,58 @@ impl ::ntex_grpc::MethodsDef for GreeterMethods {
         }
     }
 }
-#[doc = " The greeting service definition."]
+
 #[derive(Clone)]
+/// The greeting service definition.
 pub struct GreeterClient<T>(T);
+
 impl<T> GreeterClient<T> {
     #[inline]
-    #[doc = r" Create new client instance"]
+    /// Create new client instance
     pub fn new(transport: T) -> Self {
         Self(transport)
     }
 }
+
 impl<T> ::ntex_grpc::ClientInformation<T> for GreeterClient<T> {
     #[inline]
-    #[doc = r" Create new client instance"]
+    /// Create new client instance
     fn create(transport: T) -> Self {
         Self(transport)
     }
+
     #[inline]
-    #[doc = r" Get referece to underlying transport"]
+    /// Get referece to underlying transport
     fn transport(&self) -> &T {
         &self.0
     }
+
     #[inline]
-    #[doc = r" Get mut referece to underlying transport"]
+    /// Get mut referece to underlying transport
     fn transport_mut(&mut self) -> &mut T {
         &mut self.0
     }
+
     #[inline]
-    #[doc = r" Consume client and return inner transport"]
+    /// Consume client and return inner transport
     fn into_inner(self) -> T {
         self.0
     }
 }
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct GreeterSayHelloMethod;
+
 impl ::ntex_grpc::MethodDef for GreeterSayHelloMethod {
     const NAME: &'static str = "SayHello";
-    const PATH: ::ntex_grpc::types::ByteString =
-        ::ntex_grpc::types::ByteString::from_static("/helloworld.Greeter/SayHello");
+    const PATH: ::ntex_grpc::ByteString =
+        ::ntex_grpc::ByteString::from_static("/helloworld.Greeter/SayHello");
     type Input = HelloRequest;
     type Output = HelloReply;
 }
+
 impl<T: ::ntex_grpc::Transport<GreeterSayHelloMethod>> GreeterClient<T> {
-    #[doc = " Sends a greeting"]
+    /// Sends a greeting
     pub fn say_hello<'a>(
         &'a self,
         req: &'a HelloRequest,