+++ ntex-grpc-codegen/src/generator.rs
@@ -1,6 +1,4 @@
 use ntex_prost_build::{Method, Service, ServiceGenerator};
-use proc_macro2::TokenStream;
-use quote::quote;
 
 #[derive(Debug, Copy, Clone)]
 pub(crate) struct GrpcServiceGenerator;
@@ -19,144 +17,181 @@ impl ServiceGenerator for GrpcServiceGenerator {
 }
 
 fn generate_client(service: &Service, buf: &mut String) {
-    let service_ident = quote::format_ident!("{}", service.name);
-    let client_ident = quote::format_ident!("{}Client", service.name);
+    let service_ident = service.name.to_string();
+    let client_ident = format!("{}Client", service.name);
     let service_name = format!("{}.{}", service.package, service.proto_name);
-    let service_methods_name = quote::format_ident!("{}Methods", service.name);
+    let service_methods_name = format!("{}Methods", service.name);
+
     let service_methods: Vec<_> = service
         .methods
         .iter()
         .map(|m| {
-            let name = quote::format_ident!("{}", m.proto_name);
-            let m_name = quote::format_ident!("{}{}Method", service.proto_name, m.proto_name);
-            quote! {
-                #name(#m_name)
-            }
+            format!(
+                "{}({}{}Method),",
+                m.proto_name, service.proto_name, m.proto_name
+            )
         })
         .collect();
+    let service_methods = service_methods.join("\n");
+
     let mut service_methods_match: Vec<_> = service
         .methods
         .iter()
         .map(|m| {
-            let name = quote::format_ident!("{}", m.proto_name);
-            let m_name = quote::format_ident!("{}{}Method", service.proto_name, m.proto_name);
-            quote! {
-                #m_name::NAME => Some(#service_methods_name::#name(#m_name))
-            }
+            let name = m.proto_name.to_string();
+            let m_name = format!("{}{}Method", service.proto_name, m.proto_name);
+            format!(
+                "{}::NAME => Some({}::{}({})),",
+                m_name, service_methods_name, name, m_name
+            )
         })
         .collect();
-    service_methods_match.push(quote!(_ => None));
+    service_methods_match.push("_ => None".to_string());
+    let service_methods_match = service_methods_match.join("\n");
 
     let methods: Vec<_> = service
         .methods
         .iter()
         .map(|m| gen_method(m, service))
         .collect();
-    let mut comments = service.comments.leading.clone();
-    if comments.is_empty() {
-        comments.push("".to_string());
-    }
+    let methods = methods.join("\n\n");
 
-    let stream = quote! {
+    let comments: Vec<_> = service
+        .comments
+        .leading
+        .clone()
+        .into_iter()
+        .map(|s| format!("///{}", s))
+        .collect();
+    let comments = comments.join("");
 
-        pub struct #service_ident;
+    let stream = format!(
+        "pub struct {};
 
-        impl ::ntex_grpc::ServiceDef for #service_ident {
-            const NAME: &'static str = #service_name;
+        impl ::ntex_grpc::ServiceDef for {} {{
+            const NAME: &'static str = \"{}\";
+            type Methods = {};
+        }}
 
-            type Methods = #service_methods_name;
-        }
+        pub enum {} {{
+            {}
+        }}
 
-        pub enum #service_methods_name {
-            #(#service_methods),*
-        }
-        impl ::ntex_grpc::MethodsDef for #service_methods_name {
+        impl ::ntex_grpc::MethodsDef for {} {{
             #[inline]
-            fn by_name(name: &str) -> Option<Self> {
+            fn by_name(name: &str) -> Option<Self> {{
                 use ::ntex_grpc::MethodDef;
+                match name {{
+                    {}
+                }}
+            }}
+        }}
 
-                match name {
-                    #(#service_methods_match),*
-                }
-            }
-        }
-
-        #[doc = #(#comments)*]
         #[derive(Clone)]
-        pub struct #client_ident<T>(T);
+        {}
+        pub struct {}<T>(T);
 
-        impl<T> #client_ident<T> {
+        impl<T> {}<T> {{
             #[inline]
             /// Create new client instance
-            pub fn new(transport: T) -> Self {
+            pub fn new(transport: T) -> Self {{
                 Self(transport)
-            }
-        }
+            }}
+        }}
 
-        impl<T> ::ntex_grpc::ClientInformation<T> for #client_ident<T> {
+        impl<T> ::ntex_grpc::ClientInformation<T> for {}<T> {{
             #[inline]
             /// Create new client instance
-            fn create(transport: T) -> Self {
+            fn create(transport: T) -> Self {{
                 Self(transport)
-            }
+            }}
 
             #[inline]
             /// Get referece to underlying transport
-            fn transport(&self) -> &T {
+            fn transport(&self) -> &T {{
                 &self.0
-            }
+            }}
 
             #[inline]
             /// Get mut referece to underlying transport
-            fn transport_mut(&mut self) -> &mut T {
+            fn transport_mut(&mut self) -> &mut T {{
                 &mut self.0
-            }
+            }}
 
             #[inline]
             /// Consume client and return inner transport
-            fn into_inner(self) -> T {
+            fn into_inner(self) -> T {{
                 self.0
-            }
-        }
+            }}
+        }}
 
-        #(#methods)*
-    };
-    buf.push_str(&format!("{}", stream));
+        {}",
+        service_ident,
+        service_ident,
+        service_name,
+        service_methods_name,
+        service_methods_name,
+        service_methods,
+        service_methods_name,
+        service_methods_match,
+        comments,
+        client_ident,
+        client_ident,
+        client_ident,
+        methods,
+    );
+    buf.push_str(&stream);
 }
 
-fn gen_method(method: &Method, service: &Service) -> TokenStream {
+fn gen_method(method: &Method, service: &Service) -> String {
     let proto_name = &method.proto_name;
     let path = format!(
         "/{}.{}/{}",
         service.package, service.proto_name, method.proto_name
     );
 
-    let service_ident = quote::format_ident!("{}Client", service.name);
-    let method_ident = quote::format_ident!("{}", method.name);
-    let def_ident = quote::format_ident!("{}{}Method", service.name, method.proto_name);
-    let input_type = quote::format_ident!("{}", method.input_type);
-    let output_type = quote::format_ident!("{}", method.output_type);
-    let mut comments = method.comments.leading.clone();
-    if comments.is_empty() {
-        comments.push("".to_string());
-    }
+    let service_ident = format!("{}Client", service.name);
+    let method_ident = method.name.to_string();
+    let def_ident = format!("{}{}Method", service.name, method.proto_name);
+    let input_type = method.input_type.to_string();
+    let output_type = method.output_type.to_string();
+    let comments: Vec<_> = method
+        .comments
+        .leading
+        .clone()
+        .into_iter()
+        .map(|s| format!("///{}", s))
+        .collect();
+    let comments = comments.join("");
 
-    quote! {
-        #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-        pub struct #def_ident;
+    format!(
+        "#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+        pub struct {};
 
-        impl ::ntex_grpc::MethodDef for #def_ident {
-            const NAME: &'static str = #proto_name;
-            const PATH: ::ntex_grpc::types::ByteString = ::ntex_grpc::types::ByteString::from_static(#path);
-            type Input = #input_type;
-            type Output = #output_type;
-        }
+        impl ::ntex_grpc::MethodDef for {} {{
+            const NAME: &'static str = \"{}\";
+            const PATH: ::ntex_grpc::ByteString = ::ntex_grpc::ByteString::from_static(\"{}\");
+            type Input = {};
+            type Output = {};
+        }}
 
-        impl<T: ::ntex_grpc::Transport<#def_ident>> #service_ident<T> {
-            #[doc = #(#comments)*]
-            pub fn #method_ident<'a>(&'a self, req: &'a #input_type) -> ::ntex_grpc::Request<'a, T, #def_ident> {
+        impl<T: ::ntex_grpc::Transport<{}>> {}<T> {{
+            {}
+            pub fn {}<'a>(&'a self, req: &'a {}) -> ::ntex_grpc::Request<'a, T, {}> {{
                 ::ntex_grpc::Request::new(&self.0, req)
-            }
-        }
-    }
+            }}
+        }}",
+        def_ident,
+        def_ident,
+        proto_name,
+        path,
+        input_type,
+        output_type,
+        def_ident,
+        service_ident,
+        comments,
+        method_ident,
+        input_type,
+        def_ident
+    )
 }